core: scroll bars: update the gadget value while dragging
If we don't do that then during a repaint in scroll-bar/update-display the value
is initialized to the result of GADGET-VALUE and we repaint scroll bars at the
old position. This problem was masked because with full mirroring a repaint
request of the sheet (after changing the transformation) did not escalate to the
common mirror (because the sheet had its own private mirror).