ttf fonts: open the font by supplying the stream (not a filename)
This is necessary to avoid the issue where CCL assumes that streams are private
to the thread creating them. Instead of that we open the stream in "unprotected"
mode and pass it to ZPB-TTF:OPEN-FONT-LOADER.

Note that ZPB-TTF:CLOSE-FONT-LOADER does not track whether it was us who opened
the stream and it will gladly close it for us after use.

Fixes multiple frames in different threads on ccl.