core: regions: don't question the results of the newton-iteration
When computing the intersection between the ellipse and the unit circle we've
run the newton iteration and then, additionally, we've tested the result with

  (coordinate= 0 (ellipse-equation x y))

netwon-iteration gives us the best approximation but coordinate= may fail. If we
call the ellipse-intersection to determine whether there are any intersections
at all then this is faulty.