franz: use output-record instead of +flipping-ink+
`pointer-place-rubber-line*' and `pointer-input-rectangle*' now use output record
instead of flipping ink for temporary drawing.

It is better not use flipping ink because it is underspecified (e.g. from the
specification "The effect on any color other than the colors determined by those
two designs is unspecified" and also "If either design1 or design2 is not solid,
the consequences are unspecified") and it is difficult to implement it
efficiently and correctly in the modern graphic stack with color, alpha,
blending, text antialiasing ... .

For example the implementation of flipping ink in mcclim use bitwise XOR as
sugested in specification and in this way it is respected the part of
specification that say: "however, drawing the same figure twice using the same
flipping ink is guaranteed to be an "identity" operation" but not the part that
say "is defined to change the color in the background that would have been drawn
by design1 at that point into the color that would have been drawn by design2 at
that point, and vice versa".