recording: implement {x,y}-offset and remove STYLED-STRING kludge
The text line output record maintains from now on a sequence of output records
that are replayed on the line (we get rid of a kludge with STYLED-STRING class).

Moreover we implement X-OFFSET and Y-OFFSET handling in output record mapping
and replaying. This allows for working with output records in formatting macros
without changing their position (MOVE-OUTPUT-RECORD is now O(N) operation).

def-grecording defines also the REINITIALIZE-INSTANCE method that recomputes a
bounding rectangle of the output record.

Instead of a styled-string we maintain a sequence of text output records. Also
make DRAW-TEXT* bounding rectangle conform to its line (not minimal box).