cursor: maintain separately offset and extent
Previously we've maintained offset and size. This was not sufficient when
combining sizes of elements in line. Consier:

------------------------------------------------
element                | accumulated cursor
------------------------------------------------
offset = 15, size = 25 | offset = 15, size = 25
offset = 20, size = 25 | offset = 20, size = 25
------------------------------------------------

When we draw the first element, the offset (baseline) is 20, but the element
descent is 10, so the elememnt goes beyond the cursor size = 25 by 5 points.

Currently:

---------------------------------------------------
element                  | accumulated cursor
---------------------------------------------------
offset = 15, extent = 10 | offset = 15, extent = 10
offset = 20, extent = 15 | offset = 20, extent = 15

With this interpretation, the line size is correctly aggregated to 35, so all
elements fit in the line.