seos: don't assume default baseline align with default text-style
Previously we've assumed that the initial line characteristics are based on the
default stream text style. This was to ensure that empty newlines advance the
cursor, but was not conforming to the specification that says, that the first
character initializes the baseline.

This commit resets the cursor baseline to (0 0), and when the newline is issued
the cursor is updated to account for the letter "M" (without extending the line)
with the current medium text style.

Fixes #1375.