core: geometry: test for (ZEROP ANGLE) when making the rotation
This addresses only the trivial case of the numerical unstability of rotations.
For example with sheets we apply in one place transformations one after another
(instead of using the delta transformation), because otherwise the disrapency is
too big.

Also add an utility FIND-ANGLE that computes the angle in CW direction.