ttf: implement directional text rendering
We treat horizontal lines as going from left to right; that means that both
horizontal directions (and both viertical) share the same angle.

The line width is the same as the line height. We assume that there are no
glyphs that are wider than the tallest glyph; if there is, it makes no
difference, but it gives a decent default with some padding when we align to top
or bottom.